Added idea of a stack of erasures
Made a list that will store where erases happened, so when ediitng the last erased spot can be used to insert new text. NOTE TO SELF: think about using writeText function instead of repeating urself with an editText. Also, UNSURE if last erased point index is going to be correct