Pass first test of editing
Passed first test with a function that is a mix of write and erase, with the restrictions from neither carrying over. Seems to me a drawback of TDD (or at least how i am doing it), i could implement functions to pass all my tests but if previous requirements are not met i wouldnt know. I must be better aware of the tests i write and the functions the pencil fulfills. Because otherwise, editing ignores the restrictions that the pencil has when writing! Maybe time to use DRY now. This kata is masterfully designed