Fixed edge case not described in kata readme
So if you erase at the end of the page, and want to edit by adding a much longer word than you erased, previously the code would fail. But if the pencil has lead in it. then you should intuitively be able to keep going. Therefore, ediitng when there is no whitespace will append at the end. This is not shifting text, as is forbidden.